# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Matador API
  description: Matador on-demand car rental API. Customer endpoints use JWT bearer
    tokens; admin endpoints use session cookies.
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/customer/me/trips/{id}/incidents:
    post:
      tags:
      - Customer-Trips
      summary: Report incident
      description: Customer reports an incident on a trip.
      operationId: report
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerReportRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/IncidentResponse'
...
  /api/customer/bookings:
    get:
      tags:
      - Customer-Bookings
      summary: List bookings
      description: The customer's bookings.
      operationId: list_1
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageBookingResponse'
    post:
      tags:
      - Customer-Bookings
      summary: Create booking
      description: Create a booking from a quote; places a deposit hold.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/customer/bookings/{id}/cancel:
    post:
      tags:
      - Customer-Bookings
      summary: Cancel booking
      description: Cancel per the cancellation policy.
      operationId: cancel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/customer/bookings/quote:
    post:
      tags:
      - Customer-Bookings
      summary: Price quote
      description: Get a quote without creating a booking.
      operationId: quote
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QuoteResponse'
...
  /api/admin/incidents:
    get:
      tags:
      - Admin-Incidents
      summary: List incidents
      description: Filterable incidents list.
      operationId: list_6
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - OPEN
          - IN_REVIEW
          - RESOLVED
          - DISMISSED
      - name: type
        in: query
        required: false
        schema:
          type: string
          enum:
          - DAMAGE
          - OUT_OF_ZONE
          - LATE_RETURN
          - ACCIDENT
          - TICKET
          - OTHER
      - name: tripId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageIncidentResponse'
    post:
      tags:
      - Admin-Incidents
      summary: Create incident
      description: Staff creates an incident.
      operationId: create_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffCreateRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/IncidentResponse'
  /api/admin/bookings/{id}/cancel:
    post:
      tags:
      - Admin-Bookings
      summary: Cancel booking
      description: Admin-initiated cancellation.
      operationId: cancel_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/admin/bookings/{id}/assign-vehicle:
    post:
      tags:
      - Admin-Bookings
      summary: Assign vehicle
      description: Assign a specific vehicle to a booking.
      operationId: assign
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignVehicleRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/admin/bookings/{id}/activate:
    post:
      tags:
      - Admin-Bookings
      summary: Activate booking
      description: Activate a confirmed booking; starts a trip.
      operationId: activate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookingResponse'
...
  /api/admin/incidents/{id}:
    get:
      tags:
      - Admin-Incidents
      summary: Incident detail
      description: A single incident.
      operationId: get_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/IncidentResponse'
    patch:
      tags:
      - Admin-Incidents
      summary: Update incident
      description: Update status, resolution, and charges.
      operationId: update_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncidentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/IncidentResponse'
...
  /api/customer/bookings/{id}:
    get:
      tags:
      - Customer-Bookings
      summary: Booking detail
      description: A single booking owned by the customer.
      operationId: get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookingResponse'
...
  /api/admin/bookings:
    get:
      tags:
      - Admin-Bookings
      summary: List bookings
      description: Filterable bookings list.
      operationId: list_8
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - PENDING_PAYMENT
          - CONFIRMED
          - ACTIVATED
          - COMPLETED
          - CANCELLED
          - NO_SHOW
      - name: customerId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: vehicleId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageBookingResponse'
  /api/admin/bookings/{id}:
    get:
      tags:
      - Admin-Bookings
      summary: Booking detail
      description: A single booking.
      operationId: get_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookingResponse'
...
components:
  schemas:
...
    CustomerReportRequest:
      required:
      - description
      - type
      type: object
      properties:
        type:
          type: string
          enum:
          - DAMAGE
          - OUT_OF_ZONE
          - LATE_RETURN
          - ACCIDENT
          - TICKET
          - OTHER
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
    IncidentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - DAMAGE
          - OUT_OF_ZONE
          - LATE_RETURN
          - ACCIDENT
          - TICKET
          - OTHER
        severity:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
        status:
          type: string
          enum:
          - OPEN
          - IN_REVIEW
          - RESOLVED
          - DISMISSED
        chargedAmountCents:
          type: integer
          format: int64
...
    CreateBookingRequest:
      required:
      - dropoffLat
      - dropoffLng
      - pickupLat
      - pickupLng
      - quoteId
      type: object
      properties:
        quoteId:
          type: string
          format: uuid
        pickupLat:
          type: number
          format: double
        pickupLng:
          type: number
          format: double
        pickupAddress:
          type: string
        dropoffLat:
          type: number
          format: double
        dropoffLng:
          type: number
          format: double
        dropoffAddress:
          type: string
        paymentMethodId:
          type: string
          format: uuid
    BookingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingNumber:
          type: string
        status:
          type: string
          enum:
          - PENDING_PAYMENT
          - CONFIRMED
          - ACTIVATED
          - COMPLETED
          - CANCELLED
          - NO_SHOW
        vehicleClassId:
          type: string
          format: uuid
        assignedVehicleId:
          type: string
          format: uuid
        pickupAt:
          type: string
          format: date-time
        dropoffAt:
          type: string
          format: date-time
        quotedTotalCents:
          type: integer
          format: int64
        depositAmountCents:
          type: integer
          format: int64
        dropoffInZone:
          type: boolean
        depositHoldClientSecret:
          type: string
    CancelRequest:
      type: object
      properties:
        reason:
          type: string
    QuoteRequest:
      required:
      - dropoffAt
      - dropoffLat
      - dropoffLng
      - pickupAt
      - pickupLat
      - pickupLng
      - vehicleClassId
      type: object
      properties:
        vehicleClassId:
          type: string
          format: uuid
        pickupAt:
          type: string
          format: date-time
        dropoffAt:
          type: string
          format: date-time
        pickupLat:
          type: number
          format: double
        pickupLng:
          type: number
          format: double
        dropoffLat:
          type: number
          format: double
        dropoffLng:
          type: number
          format: double
        insuranceTier:
          type: string
        promoCode:
          type: string
    LineItem:
      type: object
      properties:
        type:
          type: string
        description:
          type: string
        amountCents:
          type: integer
          format: int64
    QuoteResponse:
      type: object
      properties:
        quoteId:
          type: string
          format: uuid
        dropoffInZone:
          type: boolean
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/LineItem'
        subtotalCents:
          type: integer
          format: int64
        discountCents:
          type: integer
          format: int64
        taxCents:
          type: integer
          format: int64
        totalCents:
          type: integer
          format: int64
        depositCents:
          type: integer
          format: int64
        expiresAt:
          type: string
          format: date-time
...
    StaffCreateRequest:
      required:
      - description
      - severity
      - type
      - vehicleId
      type: object
      properties:
        tripId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - DAMAGE
          - OUT_OF_ZONE
          - LATE_RETURN
          - ACCIDENT
          - TICKET
          - OTHER
        severity:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
    AssignVehicleRequest:
      required:
      - vehicleId
      type: object
      properties:
        vehicleId:
          type: string
          format: uuid
...
    UpdateIncidentRequest:
      type: object
      properties:
        status:
          type: string
          enum:
          - OPEN
          - IN_REVIEW
          - RESOLVED
          - DISMISSED
        resolutionNotes:
          type: string
        chargedAmountCents:
          type: integer
          format: int64
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
    PageBookingResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookingResponse'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
...
    PageIncidentResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/IncidentResponse'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: Customer JWT access token.
      scheme: bearer
      bearerFormat: JWT
    session-cookie:
      type: apiKey
      description: Admin session cookie.
      name: SESSION
      in: cookie